elecraft.c: Quell compiler warning of buffer size
GCC 4.8.x revealed an error in the way the destination buffer size was
calculated for storing the K3 firmware revision.  Now, instead of
declaring the array in the kenwood_priv_data structure, it is declared
as static in elecraft.c and a pointer is now assigned in the
kenwood_priv_data structure.  This allows for proper calculation of the
array size for strncpy().

(cherry picked from commit 977e92315ec322761d6742fb4fea1d95117babb9)